requirements.txt
flask
openai-whisper
torch
ffmpeg-python
TTS